caching uniforms in shader

+++ src/Graphics/GBuffer.cpp
@@ -25,7 +25,7 @@ namespace Enjon { namespace Graphics {
 		    // - Diffuse buffer
 		    glGenTextures(1, &Textures[i]);
 		    glBindTexture(GL_TEXTURE_2D, Textures[i]);
-		    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB32F, Width, Height, 0, GL_RGB, GL_FLOAT, NULL);
+		    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB16F, Width, Height, 0, GL_RGB, GL_FLOAT, NULL);
 		    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
 		    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
 			glTexParameterf(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
